requirements.txt
# ClawGeeks Platform Dependencies

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP client
httpx>=0.26.0

# Database (PostgreSQL)
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# Billing
stripe>=7.0.0

# Security
bcrypt>=4.1.0
PyJWT>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.19.0
httpx  # for TestClient

# Migrations
alembic>=1.13.0

# Utils
python-dotenv>=1.0.0
